Ignoring when both encodings are the same (#207)
Ignoring the chatting when both encodings are the same

I have the same problem issue https://github.com/Webklex/laravel-imap/issues/203 and the problem is because the source and destination encoding are the same and do not need conversion. When this happens it is returned false instead of returning the text.